agents/startup-fundraising-agent/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
python-dotenv>=1.0
httpx>=0.27
firebase-admin>=6.5.0

